fix(earn): split recommended section and fix layouts
(cherry picked from commit dc55a38fc3e1deece447cb220182ccfaab348a02)
